Fix moving average.

+++ moving-average/src/MovingAvg.h
@@ -1,14 +1,5 @@
 #pragma once
 
-
-// FIXME! Some hints about performance! 
-// Here you are using a lot of floating ops, which are always slower than fixed point integer operations
-// For example m_sum += initalValue will be relativly slow. 
-// Additionally you are storing it in a float[WINDOW_SIZE] array, which is also a bit overkill
-// Here is what i would change!
-// store m_sum, m_readings in uint16_t (halfs memory) and perform only one floating division
-// on getAvg(), this might be a bit better, the only downside is that it will have to convert a 
-// integer to a float here, but i think that's probably still faster.
 template <unsigned long long WINDOW_SIZE> class MovingAvg {
 public:
   MovingAvg(float initialValue = 0) : m_sum(0), m_next(0) {
@@ -26,7 +17,7 @@ public:
       m_next = 0;
   }
 
-  float getAvg() { return m_sum / WINDOW_SIZE; };
+  float getAvg() { return (float) m_sum / WINDOW_SIZE; };
 
   void reset(float initalValue = 0) {
     m_sum = initalValue * WINDOW_SIZE;
@@ -39,7 +30,7 @@ public:
   float *getReadings() { return m_readings; }
 
 private:
-  float m_sum;                   // sum of the m_readings array
-  int m_next;                    // index to the next reading
-  float m_readings[WINDOW_SIZE]; // interval array
+  uint16_t m_sum;                   // sum of the m_readings array
+  int16_t m_next;                    // index to the next reading
+  uint16_t m_readings[WINDOW_SIZE]; // interval array
 };
